First 2021 release working with CommonTasks 0.3.163 (#96)
* Updated readme

* Updated Azure DevOps Server image name to 2019 1.1

* Some more content

* Fixed typos

* Fix typo in Tasks\TestDscResources.ps1

* Notepad will be installed in the customization script

* Issue deploying DSCHost01 in Azure. Removing the machine for now.

* Adding Notepad++

* Fixes serialization issue, or worked around it.

* Updated DSC Resource versions

* Updated ComputerManagementDsc version

* Fixing bug

* Fixing DevOps port for Azure

* Setting Pester version to 4.10.1

* Fixed issues with Azure deployments

* Alligning passwords

* Setting Pester version to 4.10.1

* Enabled HyperV host on Azure with proper role size

* Fixed issue with hyperv host

* Updating 'NetworkingDsc' version to 8.0.0

* Improving error handling

* Added new DomainController role

* Adding missing DSC resources

* Adding feature to file server role

* Adding location Tokio

* Adding new file server nodes

* Cleanup

* Adding 'Datum.InvokeCommand' handler

* Tests did not use the node filter

* Improved error handling

* Allowing build for single node without preparing compressed modules

* Adding single node build and 'SkipCompressedModulesBuild'

* Only deploying Compress Modules if the folder 'CompressModules' has content.

* Adding netlogon log in Dev environment

* Adding DscAutoJoin solution

* Adding 'CertificateDsc' to the required DSC resource list in the lab

* Renamed 'DscAutoJoin' to 'DscAutoOnboarding' and added lab install scripts

* Removing handler that does not exist yet

* Updating DSC resource versions and missing missing DSC resources

* Fixed file name

* Renaming node files instead of deleting them

* Deleting test nodes

* Setting Datum version to 0.0.38 as 0.39.0 results in 'Unable to find type [DatumProvider]'

* Code Cleanup

* Added Datum handler 'Datum.InvokeCommand' and sample data

* Updating resource and configuration module versions

* Updating module versions

* DSC Datebase is created by AutomatedLab

* Adding missing variable

* Fixed some links and typos

* Added error handling for transcription

* Added error handling in case there is only one DscResource

* Updated Windows and ADO version

* Removed unused variable

* Installing 'xDscDiagnostics' on all nodes as required by JEA endpoint

* Adding license and fixes #84

* Adding 'testResults.xml' to gitignore

* Updated module and resource versions

* Updating config module version

* Moving settings to the top

* Fixed file names

* Skipping PublishBuildArtifacts for Compressed Modules if added a new node

* Incresing DscLcmController intervals as the LCM consumes too much memory that way

* Added task output as a list for easier usage on the command line

* Added test provided by @ykuijs and error handling when calling 'Get-FilteredConfigurationData'

* Remove power shell feed credentials (#86)

* Adding a test for Windows PowerShell in Init task

* Removing credentials for contacting the NuGet feed

* Removing NuGet credential handling

* Added feature 'Send-DscTaggingData' (#87)

* Adding a test for Windows PowerShell in Init task

* Updating to latest Azure DevOps ISOs

* Added function 'Start-LabDscConfiguration' to helper module

* Fixed file name

* Added 'SendDscTaggingData' JEA endpoint and config item

* 'Send-DscTaggingData' fixes and adding endpoint creation to lab script

* Updated CommonTasks version

* Adding 'SendDscTaggingData' feature (#93)

* Adding a test for Windows PowerShell in Init task

* Updating to latest Azure DevOps ISOs

* Added function 'Start-LabDscConfiguration' to helper module

* Fixed file name

* Added 'SendDscTaggingData' JEA endpoint and config item

* 'Send-DscTaggingData' fixes and adding endpoint creation to lab script

* Updated CommonTasks version

* Updating to latest Azure DevOps ISOs

* Added function 'Start-LabDscConfiguration' to helper module

* Fixed file name

* Added 'SendDscTaggingData' JEA endpoint and config item

* 'Send-DscTaggingData' fixes and adding endpoint creation to lab script

* Updated CommonTasks version

* Added new 'AdditionalNodeInformation' report

* Added new 'NodeStatusSimple' report

* Making the node variable being propagated into the configurations

* Added error handling

* Added some logging

* Added some error handling as restarting the WinRM service interrupts the connection

* Fixes #88

* Formatting and Cleanup

* Setting the filter seems no longer needed

* Adding new requirements according to CommonTasks

* Adding debug messages in case configurations do not work

* Typos

* Added 'MaxLcmRuntime' to config according to CommonTasks build 105

* Fixing config data for 'ChocolateyPackages'

* Format

* Updated software versions

* fix parameter handling if script default parameters are changed & add link to YAML reference (#79)

* fix parameter handling if script default parameters are changed

* revert some changes

* add link to YAML reference in CommonTasks


* Updates according to updates in CommonTasks (#95)

* Adding new required modules

* Updated 'SqlServerDsc' version

* Added additional resources according to changed in CommonTasks

* Moving PSDepend files to ProjectPath for easier discovery

* Commented out debug output

* Bug fix

* Updated DSC resource version numbers

* Updated vesion numbers and using latest version of CommonTasks.

* Updated config data for 'NetworkIpConfiguration' according to updates in CommonTasks

+++ DSC/Build/Tasks/Deploy.ps1
@@ -32,7 +32,13 @@ Task Deploy {
         }
         
         Copy-DscMof -MofPath "$BuildOutput\MOF" -TargetPath $env:DscConfiguration -Environment $env:RELEASE_ENVIRONMENTNAME
-        Copy-Item -Path "$BuildOutput\CompressedModules\*" -Destination $env:DscModules
+        Copy-DscMof -MofPath "$BuildOutput\MOF" -TargetPath $env:DscConfiguration -Environment $env:RELEASE_ENVIRONMENTNAME
+        if (Test-Path -Path "$BuildOutput\CompressedModules\*") {
+            Copy-Item -Path "$BuildOutput\CompressedModules\*" -Destination $env:DscModules
+        }
+        else {
+            Write-Host "The folder '$BuildOutput\CompressedModules\*' does not exist, skipping deployment of CompressedModules." 
+        }
     }
     
 }
